factored out read_spans
